# backend/auth-admin-api/src/main/resources/META-INF/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: mja-api
  description: API für meine Benutzerverwaltung
  contact:
    name: Heike Winkelvoß
    url: https://mathe-jung-alt.de
    email: info@egladil.de
  license:
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 3.0.0
servers:
- url: http://localhost:9020
  description: Auto generated value
- url: http://0.0.0.0:9020
  description: Auto generated value
...
paths:
  /auth-admin-api/benutzer:
    post:
      tags:
      - BenutzerResource
      summary: "Gibt alle Benutzer zurück, die auf die gegebene Suchanfrage passen."
      operationId: findUsers
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BenutzerSuchparameter"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenutzerSearchResult"
        "400":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationErrorResponseDto"
        "401":
          content:
            application/json: {}
        "403":
          description: "kann auch vorkommen, wenn mod_security zuschlägt"
          content:
            application/json: {}
        "500":
          description: server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePayload"
      security:
      - SecurityScheme: []
  /auth-admin-api/benutzer/{uuid}:
    put:
      tags:
      - BenutzerResource
      summary: Setzt für den gegebenen Benutzer den neuen Aktivierungsstatus
      operationId: aktivierungsstatusAendern
      parameters:
      - name: uuid
        in: path
        description: "UUID des Benutzers, der geändert werden soll"
        required: true
        schema:
          maxLength: 36
          pattern: "^[abcdef\\d\\-]*$"
          type: string
        example: a4c4d45e-4a81-4bde-a6a3-54464801716d
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Aktivierungsstatus"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateBenutzerResponseDto"
        "400":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationErrorResponseDto"
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePayload"
        "403":
          description: "kann auch vorkommen, wenn mod_security zuschlägt"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePayload"
        "404":
          description: "kann auch vorkommen, wenn mod_security zuschlägt"
        "500":
          description: server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePayload"
      security:
      - SecurityScheme: []
    delete:
      tags:
      - BenutzerResource
      summary: Löscht das durch die uuid definierte Benutzerkonto vollständig aus
        allen Anwendungen.
      operationId: benutzerLoeschen
      parameters:
      - name: uuid
        in: path
        description: "UUID des Benutzers, der gelöscht werden soll"
        required: true
        schema:
          type: string
        example: a4c4d45e-4a81-4bde-a6a3-54464801716d
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteBenutzerResponseDto"
        "400":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationErrorResponseDto"
        "401":
          content:
            application/json: {}
        "403":
          description: "kann auch vorkommen, wenn mod_security zuschlägt"
          content:
            application/json: {}
        "500":
          description: server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePayload"
      security:
      - SecurityScheme: []
  /auth-admin-api/infomails:
    get:
      tags:
      - InfomailsResource
      summary: Gibt alle gespeicherten Infomailtexte zurück.
      operationId: loadInfomails
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InfomailResponseDto"
        "400":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationErrorResponseDto"
        "401":
          content:
            application/json: {}
        "403":
          description: "kann auch vorkommen, wenn mod_security zuschlägt"
          content:
            application/json: {}
        "500":
          description: server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePayload"
      security:
      - SecurityScheme: []
    post:
      tags:
      - InfomailsResource
      summary: Legt eine Infomail an.
      operationId: infomailAnlegen
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InfomailRequestDto"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfomailResponseDto"
        "400":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationErrorResponseDto"
        "401":
          content:
            application/json: {}
        "403":
          description: "kann auch vorkommen, wenn mod_security zuschlägt"
          content:
            application/json: {}
        "500":
          description: server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePayload"
      security:
      - SecurityScheme: []
  /auth-admin-api/infomails/{uuid}:
    put:
      tags:
      - InfomailsResource
      summary: Ändert eine Infomail an.
      operationId: infomailAendern
      parameters:
      - name: uuid
        in: path
        description: "UUID der Infomail, die geändert werden soll"
        required: true
        schema:
          maxLength: 36
          pattern: "^[abcdef\\d\\-]*$"
          type: string
        example: a4c4d45e-4a81-4bde-a6a3-54464801716d
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InfomailRequestDto"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateInfomailResponseDto"
        "400":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationErrorResponseDto"
        "401":
          content:
            application/json: {}
        "403":
          description: "kann auch vorkommen, wenn mod_security zuschlägt"
          content:
            application/json: {}
        "404":
          description: wenn es die Entity mit der uuid gar nicht gibt
          content:
            application/json: {}
        "500":
          description: server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePayload"
      security:
      - SecurityScheme: []
...
components:
  schemas:
    Aktivierungsstatus:
      description: DTO für den neuen Benutzerstatus
      type: object
      properties:
        aktiviert:
          type: boolean
...
    BenutzerSearchResult:
      description: Ergebnis der Suche
      type: object
      properties:
        anzahlGesamt:
          format: int32
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/BenutzerTrefferlisteItem"
    BenutzerSuchparameter:
      description: BenutzerSuchparameter. Gesucht wird mit und-Verknüpfung und like
      type: object
      properties:
        uuid:
          description: Teil der UUID eines Users
          maxLength: 36
          pattern: "^[abcdef\\d\\-]*$"
          type: string
          example: 732b2ed8
        vorname:
          description: Teil des Vornamens eines Users
          maxLength: 100
          pattern: "^[abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZÄäÀàÁáÂâ\
            ÃãĀāÅåĂăĄąǍǎǞǟǺǻẠạẪẫẬẢảẤấẦầẮắªẰằẲẳẴẵẶặÆæǼǽḂḃÇçĆćĊċČčÐðĎďĐđḊḋḐḑD̂d̂ÈèÉé\
            ÊêËëĒēĔĕĖėĘęĚěẼẽỄễẸẹẺẻỀềẾếỆệƏəḞḟĞğĠġĢģǤǥǦǧǴǵḠḡĦħȞȟḤḥḦḧÌìÍíÎîÏïĨĩĪīĬĭĮį\
            İıǏǐỊịỈỉĴĵǰJ̌ĶķǨǩḰḱĸĹĺĻļĽľĿŀŁłL̂l̂M̂m̂ṀṁN̂n̂ÑñŃńŅņŇňŊŋṄṅŉÖöÒòÓóÔôÕõŌōŎ\
            ŏŐőƠơǑǒǪǫǬǭǾǿȪȫȮȯȰȱỌọỎỏỖỗºỒồỐốỘộỜờỚớØøŒœṖṗŔŕŖŗŘřŚśŞşŠšȘșṠṡṢṣßẞŢţŤťŦŧȚț\
            ṪṫÞþÜüÙùÚúÛûŨũŪūŮůŰűŲųƯưǓǔỤụỦủỨứỪừỬửỮữỰựŴŵẀẁẂẃẄẅẌẍÝýŸÿŶŷȲȳẎẏỲỳỸỹỴỵỶỷƷʒ\
            ǮǯŹźŻżŽžẐẑẒẓ0123456789\" \\-_\\.,'`'\\@()/‘+:;=\\[\\]{}!#$§%&\\*\\?\\\\\
            ^|~°]*$"
          type: string
          example: Rainer
        nachname:
          description: Teil des Nachnamens eines Users
          maxLength: 100
          pattern: "^[abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZÄäÀàÁáÂâ\
            ÃãĀāÅåĂăĄąǍǎǞǟǺǻẠạẪẫẬẢảẤấẦầẮắªẰằẲẳẴẵẶặÆæǼǽḂḃÇçĆćĊċČčÐðĎďĐđḊḋḐḑD̂d̂ÈèÉé\
            ÊêËëĒēĔĕĖėĘęĚěẼẽỄễẸẹẺẻỀềẾếỆệƏəḞḟĞğĠġĢģǤǥǦǧǴǵḠḡĦħȞȟḤḥḦḧÌìÍíÎîÏïĨĩĪīĬĭĮį\
            İıǏǐỊịỈỉĴĵǰJ̌ĶķǨǩḰḱĸĹĺĻļĽľĿŀŁłL̂l̂M̂m̂ṀṁN̂n̂ÑñŃńŅņŇňŊŋṄṅŉÖöÒòÓóÔôÕõŌōŎ\
            ŏŐőƠơǑǒǪǫǬǭǾǿȪȫȮȯȰȱỌọỎỏỖỗºỒồỐốỘộỜờỚớØøŒœṖṗŔŕŖŗŘřŚśŞşŠšȘșṠṡṢṣßẞŢţŤťŦŧȚț\
            ṪṫÞþÜüÙùÚúÛûŨũŪūŮůŰűŲųƯưǓǔỤụỦủỨứỪừỬửỮữỰựŴŵẀẁẂẃẄẅẌẍÝýŸÿŶŷȲȳẎẏỲỳỸỹỴỵỶỷƷʒ\
            ǮǯŹźŻżŽžẐẑẒẓ0123456789\" \\-_\\.,'`'\\@()/‘+:;=\\[\\]{}!#$§%&\\*\\?\\\\\
            ^|~°]*$"
          type: string
          example: Honig
        email:
          description: Teil der email eines Users
          maxLength: 255
          pattern: "^[a-zA-Z0-9\\.!#$%&'*+/=\\?\\^_`{|}~\\-@]*$"
          type: string
          example: egladil
        rolle:
          description: Teil der Rolle eines Users.
          maxLength: 150
          pattern: "^[A-Za-z_,]*$"
          type: string
          example: STANDARD
        dateModified:
          description: "Teil des Datums, an dem die Daten das letzte Mal geändert\
            \ wurden - meist letztes Login Format '2019-09-14 18:40:06'"
          maxLength: 19
          pattern: "^[\\d\\-.:]*$"
          type: string
          example: "2021"
        sortByLabelname:
          description: "Name des Feldes, nach dem sortiert werden soll. Der Name muss\
            \ sich auf die enum BenutzerSortColumn abbilden lassen"
          maxLength: 100
          pattern: "^[A-Za-zäöüßÄÖÜ\\-_ ]*$"
          type: string
          example: vorname
        pageIndex:
          format: int32
          description: "Pagination: welche Seite wird abgefragt"
          default: 0
          type: integer
          example: 2
        pageSize:
          format: int32
          description: "Pagination: wievile Treffer werden abgefragt"
          default: 25
          type: integer
          example: 25
        sortDirection:
          description: Sortierreihenfolge (asc oder desc)
          type: string
          allOf:
          - $ref: "#/components/schemas/SortDirection"
          example: asc
    BenutzerTrefferlisteItem:
      description: "Ein Elememt der Trefferliste. Es sind die Attribute, nach denen\
        \ auch gesucht werden kann."
      type: object
      properties:
        uuid:
          description: UUID des Users
          maxLength: 36
          pattern: "^[abcdef\\d\\-]*$"
          type: string
          example: 732b2ed8-b9b7-4800-8685-38fd09d330cf
        vorname:
          description: Vorname eines Users
          maxLength: 100
          pattern: "^[abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZÄäÀàÁáÂâ\
            ÃãĀāÅåĂăĄąǍǎǞǟǺǻẠạẪẫẬẢảẤấẦầẮắªẰằẲẳẴẵẶặÆæǼǽḂḃÇçĆćĊċČčÐðĎďĐđḊḋḐḑD̂d̂ÈèÉé\
            ÊêËëĒēĔĕĖėĘęĚěẼẽỄễẸẹẺẻỀềẾếỆệƏəḞḟĞğĠġĢģǤǥǦǧǴǵḠḡĦħȞȟḤḥḦḧÌìÍíÎîÏïĨĩĪīĬĭĮį\
            İıǏǐỊịỈỉĴĵǰJ̌ĶķǨǩḰḱĸĹĺĻļĽľĿŀŁłL̂l̂M̂m̂ṀṁN̂n̂ÑñŃńŅņŇňŊŋṄṅŉÖöÒòÓóÔôÕõŌōŎ\
            ŏŐőƠơǑǒǪǫǬǭǾǿȪȫȮȯȰȱỌọỎỏỖỗºỒồỐốỘộỜờỚớØøŒœṖṗŔŕŖŗŘřŚśŞşŠšȘșṠṡṢṣßẞŢţŤťŦŧȚț\
            ṪṫÞþÜüÙùÚúÛûŨũŪūŮůŰűŲųƯưǓǔỤụỦủỨứỪừỬửỮữỰựŴŵẀẁẂẃẄẅẌẍÝýŸÿŶŷȲȳẎẏỲỳỸỹỴỵỶỷƷʒ\
            ǮǯŹźŻżŽžẐẑẒẓ0123456789\" \\-_\\.,'`'\\@()/‘+:;=\\[\\]{}!#$§%&\\*\\?\\\\\
            ^|~°]*$"
          type: string
          example: Rainer
        nachname:
          description: ' Nachname eines Users'
          maxLength: 100
          pattern: "^[abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZÄäÀàÁáÂâ\
            ÃãĀāÅåĂăĄąǍǎǞǟǺǻẠạẪẫẬẢảẤấẦầẮắªẰằẲẳẴẵẶặÆæǼǽḂḃÇçĆćĊċČčÐðĎďĐđḊḋḐḑD̂d̂ÈèÉé\
            ÊêËëĒēĔĕĖėĘęĚěẼẽỄễẸẹẺẻỀềẾếỆệƏəḞḟĞğĠġĢģǤǥǦǧǴǵḠḡĦħȞȟḤḥḦḧÌìÍíÎîÏïĨĩĪīĬĭĮį\
            İıǏǐỊịỈỉĴĵǰJ̌ĶķǨǩḰḱĸĹĺĻļĽľĿŀŁłL̂l̂M̂m̂ṀṁN̂n̂ÑñŃńŅņŇňŊŋṄṅŉÖöÒòÓóÔôÕõŌōŎ\
            ŏŐőƠơǑǒǪǫǬǭǾǿȪȫȮȯȰȱỌọỎỏỖỗºỒồỐốỘộỜờỚớØøŒœṖṗŔŕŖŗŘřŚśŞşŠšȘșṠṡṢṣßẞŢţŤťŦŧȚț\
            ṪṫÞþÜüÙùÚúÛûŨũŪūŮůŰűŲųƯưǓǔỤụỦủỨứỪừỬửỮữỰựŴŵẀẁẂẃẄẅẌẍÝýŸÿŶŷȲȳẎẏỲỳỸỹỴỵỶỷƷʒ\
            ǮǯŹźŻżŽžẐẑẒẓ0123456789\" \\-_\\.,'`'\\@()/‘+:;=\\[\\]{}!#$§%&\\*\\?\\\\\
            ^|~°]*$"
          type: string
          example: Honig
        email:
          description: email eines Users
          maxLength: 255
          pattern: "^[a-zA-Z0-9\\.!#$%&'*+/=\\?\\^_`{|}~\\-@]*$"
          type: string
          example: test@egladil.de
        rollen:
          description: Rollen eines Users.
          maxLength: 150
          pattern: "^[A-Z_,]*$"
          type: string
          example: STANDARD
        aktiviert:
          description: "Flag ,ob der user aktiviert ist"
          type: boolean
          example: false
        dateModified:
          description: "Datum, an dem die Daten das letzte Mal geändert wurden - meist\
            \ letztes Login"
          maxLength: 19
          pattern: "^[\\d.\\-:]*$"
          type: string
          example: 2019-09-14 18:40:06
    DeleteBenutzerResponseDto:
      description: Response-Objekt für das Löschen eines Benutzers
      type: object
      properties:
        uuid:
          description: UUID des Users
          maxLength: 36
          pattern: "^[abcdef\\d\\-]*$"
          type: string
          example: 732b2ed8-b9b7-4800-8685-38fd09d330cf
    InfomailRequestDto:
      description: Betreff und Text für eine Infomail
      type: object
      properties:
        betreff:
          maxLength: 100
          pattern: "^[abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZÄäÀàÁáÂâ\
            ÃãĀāÅåĂăĄąǍǎǞǟǺǻẠạẪẫẬẢảẤấẦầẮắªẰằẲẳẴẵẶặÆæǼǽḂḃÇçĆćĊċČčÐðĎďĐđḊḋḐḑD̂d̂ÈèÉé\
            ÊêËëĒēĔĕĖėĘęĚěẼẽỄễẸẹẺẻỀềẾếỆệƏəḞḟĞğĠġĢģǤǥǦǧǴǵḠḡĦħȞȟḤḥḦḧÌìÍíÎîÏïĨĩĪīĬĭĮį\
            İıǏǐỊịỈỉĴĵǰJ̌ĶķǨǩḰḱĸĹĺĻļĽľĿŀŁłL̂l̂M̂m̂ṀṁN̂n̂ÑñŃńŅņŇňŊŋṄṅŉÖöÒòÓóÔôÕõŌōŎ\
            ŏŐőƠơǑǒǪǫǬǭǾǿȪȫȮȯȰȱỌọỎỏỖỗºỒồỐốỘộỜờỚớØøŒœṖṗŔŕŖŗŘřŚśŞşŠšȘșṠṡṢṣßẞŢţŤťŦŧȚț\
            ṪṫÞþÜüÙùÚúÛûŨũŪūŮůŰűŲųƯưǓǔỤụỦủỨứỪừỬửỮữỰựŴŵẀẁẂẃẄẅẌẍÝýŸÿŶŷȲȳẎẏỲỳỸỹỴỵỶỷƷʒ\
            ǮǯŹźŻżŽžẐẑẒẓ0123456789\" \\-_\\.,'`'\\@()/‘+:;=\\[\\]{}!#$§%&\\*\\?\\\\\
            ^|~°]*$"
          type: string
        mailtext:
          maxLength: 65535
          pattern: "^[abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZÄäÀàÁáÂâ\
            ÃãĀāÅåĂăĄąǍǎǞǟǺǻẠạẪẫẬẢảẤấẦầẮắªẰằẲẳẴẵẶặÆæǼǽḂḃÇçĆćĊċČčÐðĎďĐđḊḋḐḑD̂d̂ÈèÉé\
            ÊêËëĒēĔĕĖėĘęĚěẼẽỄễẸẹẺẻỀềẾếỆệƏəḞḟĞğĠġĢģǤǥǦǧǴǵḠḡĦħȞȟḤḥḦḧÌìÍíÎîÏïĨĩĪīĬĭĮį\
            İıǏǐỊịỈỉĴĵǰJ̌ĶķǨǩḰḱĸĹĺĻļĽľĿŀŁłL̂l̂M̂m̂ṀṁN̂n̂ÑñŃńŅņŇňŊŋṄṅŉÖöÒòÓóÔôÕõŌōŎ\
            ŏŐőƠơǑǒǪǫǬǭǾǿȪȫȮȯȰȱỌọỎỏỖỗºỒồỐốỘộỜờỚớØøŒœṖṗŔŕŖŗŘřŚśŞşŠšȘșṠṡṢṣßẞŢţŤťŦŧȚț\
            ṪṫÞþÜüÙùÚúÛûŨũŪūŮůŰűŲųƯưǓǔỤụỦủỨứỪừỬửỮữỰựŴŵẀẁẂẃẄẅẌẍÝýŸÿŶŷȲȳẎẏỲỳỸỹỴỵỶỷƷʒ\
            ǮǯŹźŻżŽžẐẑẒẓ0123456789\" \\-_\\.,'`'\\@()/‘+:;=\\[\\]{}!#$§%&\\*\\?\\\\\
            ^|~°]*$"
          type: string
    InfomailResponseDto:
      description: eine Infomail
      type: object
      properties:
        uuid:
          description: UUID des Infomailtextes
          maxLength: 36
          pattern: "^[abcdef\\d\\-]*$"
          type: string
          example: 5eaaf0ed-4949-4ccd-a9c3-2db9af3559c2
        betreff:
          description: "Betreff der Mail, die versendet weren wird"
          maxLength: 100
          pattern: "^[abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZÄäÀàÁáÂâ\
            ÃãĀāÅåĂăĄąǍǎǞǟǺǻẠạẪẫẬẢảẤấẦầẮắªẰằẲẳẴẵẶặÆæǼǽḂḃÇçĆćĊċČčÐðĎďĐđḊḋḐḑD̂d̂ÈèÉé\
            ÊêËëĒēĔĕĖėĘęĚěẼẽỄễẸẹẺẻỀềẾếỆệƏəḞḟĞğĠġĢģǤǥǦǧǴǵḠḡĦħȞȟḤḥḦḧÌìÍíÎîÏïĨĩĪīĬĭĮį\
            İıǏǐỊịỈỉĴĵǰJ̌ĶķǨǩḰḱĸĹĺĻļĽľĿŀŁłL̂l̂M̂m̂ṀṁN̂n̂ÑñŃńŅņŇňŊŋṄṅŉÖöÒòÓóÔôÕõŌōŎ\
            ŏŐőƠơǑǒǪǫǬǭǾǿȪȫȮȯȰȱỌọỎỏỖỗºỒồỐốỘộỜờỚớØøŒœṖṗŔŕŖŗŘřŚśŞşŠšȘșṠṡṢṣßẞŢţŤťŦŧȚț\
            ṪṫÞþÜüÙùÚúÛûŨũŪūŮůŰűŲųƯưǓǔỤụỦủỨứỪừỬửỮữỰựŴŵẀẁẂẃẄẅẌẍÝýŸÿŶŷȲȳẎẏỲỳỸỹỴỵỶỷƷʒ\
            ǮǯŹźŻżŽžẐẑẒẓ0123456789\" \\-_\\.,'`'\\@()/‘+:;=\\[\\]{}!#$§%&\\*\\?\\\\\
            ^|~°]*$"
          type: string
        mailtext:
          description: "Text der Mail, die versendet werden wird"
          maxLength: 65535
          pattern: "^[abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZÄäÀàÁáÂâ\
            ÃãĀāÅåĂăĄąǍǎǞǟǺǻẠạẪẫẬẢảẤấẦầẮắªẰằẲẳẴẵẶặÆæǼǽḂḃÇçĆćĊċČčÐðĎďĐđḊḋḐḑD̂d̂ÈèÉé\
            ÊêËëĒēĔĕĖėĘęĚěẼẽỄễẸẹẺẻỀềẾếỆệƏəḞḟĞğĠġĢģǤǥǦǧǴǵḠḡĦħȞȟḤḥḦḧÌìÍíÎîÏïĨĩĪīĬĭĮį\
            İıǏǐỊịỈỉĴĵǰJ̌ĶķǨǩḰḱĸĹĺĻļĽľĿŀŁłL̂l̂M̂m̂ṀṁN̂n̂ÑñŃńŅņŇňŊŋṄṅŉÖöÒòÓóÔôÕõŌōŎ\
            ŏŐőƠơǑǒǪǫǬǭǾǿȪȫȮȯȰȱỌọỎỏỖỗºỒồỐốỘộỜờỚớØøŒœṖṗŔŕŖŗŘřŚśŞşŠšȘșṠṡṢṣßẞŢţŤťŦŧȚț\
            ṪṫÞþÜüÙùÚúÛûŨũŪūŮůŰűŲųƯưǓǔỤụỦủỨứỪừỬửỮữỰựŴŵẀẁẂẃẄẅẌẍÝýŸÿŶŷȲȳẎẏỲỳỸỹỴỵỶỷƷʒ\
            ǮǯŹźŻżŽžẐẑẒẓ0123456789\" \\-_\\.,'`'\\@()/‘+:;=\\[\\]{}!#$§%&\\*\\?\\\\\
            ^|~°]*$"
          type: string
        uuidsMailversandauftraege:
          description: UUIDs der Versandaufträge mit diesem oder einem angepassten
            Mailtext
          type: array
          items:
            type: string
...
    MessagePayload:
      description: eine kategorisierte Meldung
      type: object
      properties:
        level:
          description: "Level der Message: INFO | WARN | ERROR"
          type: string
        message:
          description: die message
          type: string
...
    SortDirection:
      enum:
      - asc
      - desc
      type: string
    UpdateBenutzerResponseDto:
      description: "Ergebnis nach dem Ändern des Benutzers. Attribut benutzer wird\
        \ null sein, wenn statusCode != 200"
      type: object
      properties:
        uuid:
          type: string
        benuzer:
          $ref: "#/components/schemas/BenutzerTrefferlisteItem"
    UpdateInfomailResponseDto:
      description: "Ergebnis nach dem Ändern des InfomailTexts. Attribut infomail\
        \ wird null sein, wenn statusCode != 200"
      type: object
      properties:
        uuid:
          type: string
        infomail:
          $ref: "#/components/schemas/InfomailResponseDto"
    ValidationErrorResponseDto:
      type: object
      properties:
        fieldName:
          type: string
        message:
          type: string
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT